+++ src/main/java/com/example/searchengine_ver1/backendapi/config/SearchConfig.java
@@ -1,0 +1,21 @@
+package com.example.searchengine_ver1.backendapi.config;
+
+import com.example.searchengine_ver1.backendapi.proxy.CachedSearchProxy;
+import com.example.searchengine_ver1.backendapi.proxy.SearchServiceInterface;
+import com.example.searchengine_ver1.backendapi.service.RealSearchService;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+/**
+ *  Wire CachedSearchProxy with RealSearchProxy through Spring boot framework
+ * */
+@Configuration
+public class SearchConfig {
+
+    @Bean
+    public SearchServiceInterface searchServiceInterface(RealSearchService realSearchService) {
+        return new CachedSearchProxy(realSearchService);
+    }
+}
+