+++ src/main/java/com/example/searchengine_ver1/backendapi/controller/SearchController.java
@@ -1,12 +1,15 @@
 package com.example.searchengine_ver1.backendapi.controller;
 
-import com.example.searchengine_ver1.backendapi.service.SearchService;
+import com.example.searchengine_ver1.backendapi.proxy.SearchServiceInterface;
+import com.example.searchengine_ver1.backendapi.service.RealSearchService;
+import com.example.searchengine_ver1.core.analysis.MetadataAnalyzer;
 import com.example.searchengine_ver1.core.model.FileIndex;
 import com.example.searchengine_ver1.core.model.SearchResponse;
 import com.example.searchengine_ver1.exception.ContentNotPresentException;
 import com.example.searchengine_ver1.core.widget.WidgetFactory;
 import com.example.searchengine_ver1.core.widget.Widget;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,15 +20,19 @@ import java.util.List;
 public class SearchController {
 
     @Autowired
-    private SearchService searchService;
+    private RealSearchService realSearchService;
     private final WidgetFactory widgetFactory = new WidgetFactory();
+    private final MetadataAnalyzer metadataAnalyzer = new MetadataAnalyzer();
 
+    @Autowired
+    @Qualifier("cachedSearchProxy")
+    private SearchServiceInterface searchService;
     @GetMapping
     public ResponseEntity<SearchResponse> search(@RequestParam String query) {
         List<FileIndex> results;
 
         try {
-            results = searchService.search(query);
+            results = realSearchService.search(query);
         } catch (ContentNotPresentException e) {
             return ResponseEntity.badRequest().body(new SearchResponse("Please introduce content."));
         }
@@ -41,7 +48,7 @@ public class SearchController {
         response.setResults(results);
         response.setSpecialWidgets(specialWidgets);
         response.setContextWidgets(contextWidgets);
-        //response.setMetadataSummary(metadataAnalyzer.analyze(results));
+        response.setMetadataSummary(metadataAnalyzer.analyze(results));
         response.setMessage("Success");
 
         return ResponseEntity.ok(response);